+++ src/neural_network/layer/global_average_pooling_2d.rs
@@ -157,5 +157,5 @@ impl Layer for GlobalAveragePooling2D {
         "GlobalAveragePooling2D"
     }
 
-    layer_functions_avg_pooling!();
+    layer_functions_global_pooling!();
 }