+++ src/neural_network/layer/global_max_pooling_1d.rs
@@ -158,5 +158,5 @@ impl Layer for GlobalMaxPooling1D {
         "GlobalMaxPooling1D"
     }
 
-    layer_functions_avg_pooling!();
+    layer_functions_global_pooling!();
 }