+++ src/neural_network/layer/global_max_pooling_2d.rs
@@ -180,5 +180,5 @@ impl Layer for GlobalMaxPooling2D {
         "GlobalMaxPooling2D"
     }
 
-    layer_functions_avg_pooling!();
+    layer_functions_global_pooling!();
 }